+++ click_extra/pytest.py
@@ -334,7 +334,7 @@ default_debug_colored_verbose_log = (
 
 default_config_file_pattern = (
     r"\*\.toml\|\*\.yaml\|\*\.yml\|\*\.json\|\*\.json5\|\*\.jsonc\|\*\.hjson\|\*\.ini\|"
-    r"\*\.xml"
+    r"\*\.xml\|pyproject\.toml"
 )
 default_debug_uncolored_config = (
     rf"debug: Load configuration matching .+{default_config_file_pattern}\n"